+++ src/osd/mod.rs
@@ -3,10 +3,6 @@ mod fc_firmware;
 mod frame;
 mod glyph;
 mod osd_file;
-mod preview;
 
 pub use frame::Frame;
 pub use osd_file::OsdFile;
-pub use preview::calculate_horizontal_offset;
-pub use preview::calculate_vertical_offset;
-pub use preview::osd_preview;