+++ src/srt/mod.rs
@@ -1,0 +1,7 @@
+mod error;
+mod frame;
+mod srt_file;
+
+pub use frame::SrtFrame;
+pub use frame::SrtFrameData;
+pub use srt_file::SrtFile;