+++ src/ui/mod.rs
@@ -1,10 +1,13 @@
 mod app;
 mod bottom_panel;
 mod central_panel;
+mod osd_preview;
 mod render_status;
 mod side_panel;
 mod top_panel;
 mod utils;
 
+pub use app::OsdOptions;
+pub use app::SrtOptions;
 pub use app::WalksnailOsdTool;
 pub use render_status::RenderStatus;